Update to latest components, PHP 8.1 compat

+++ src/DataTransferObjects/Schemas/Shipping/CurrencySchema.php
@@ -7,7 +7,7 @@ use Spatie\DataTransferObject\DataTransferObject;
 
 class CurrencySchema extends DataTransferObject
 {
-    public int $value;
+    public float $value;
 
     #[MaxLengthValidator(3)]
     public string $unit;
